requirements.txt
fastapi
fastapi_utils
uvicorn
passlib
pyjwt
sqlalchemy
pandas
python-multipart
cryptography
alembic
email-validator
password-validator
phonenumbers
apscheduler
pydantic
asyncpg
bcrypt
python-dotenv
openpyxl
redis
aioredis